LayoutTests/fast/events/event-leak-objects-expected.txt

PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS successfullyParsed is true

TEST COMPLETE
